Update sketch.js
The timing keeps changing slightly so the location of the tennis ball keeps changing and isn't always in the pipe. I wanted the targets to affect the path of the tennis ball and was able to get that to work for the first 3 targets but the only way it works is if you click and hold over the target from the start until the ball reaches the end. Then at the end, something happens depending on which box the ball ends up in. I also wanted the bullet to shoot out of the gun using position, velocity, and acceleration but it didnt quite work so I left it as if the mouse was clicked, something appears in the aim.